+++ CHANGELOG.md
@@ -8,12 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - option to update `minimumStakingTokenPercentage`.
 - `initialBlockHeights` to set the initial block-height per shard
+- fetch VerifierRegistry contract from ContractRegistry
 
 ### Changed
 - smart contracts are complient with solidity 0.5
 - unit tests are complient with web3 1.0
 - when verifier proposing, `Chain` checks, if verifier is active instead
   of if it is only created.
+- split Chain contract into two separate contracts (storage and manager)
 
 ## [0.3.0] - 2018-12-13
 ### Added: